requirements.txt
streamlit>=1.32.0
openai>=1.40.0
python-dotenv>=1.0.1
numpy>=1.26.0
faiss-cpu>=1.8.0
pymupdf>=1.24.0
